Integrate Lua script to game loop
Integrate Lua script to OpenGL game loop. Made a sample run to let Lua eneity change the clear color in screen

+++ source/LuaHeaderTool/SyntaxAnalyzer.cpp
@@ -48,7 +48,7 @@ void SyntaxAnalyzer::Run(const std::vector<HeaderTokenizer::Token>& tokens, std:
 	mInput = &tokens;
 	mOutput = &output;
 
-	for (int i = 0; i < tokens.size(); ++i)
+	for (size_t i = 0; i < tokens.size(); ++i)
 	{
 		auto func = sStateFunctions[static_cast<size_t>(mState)];
 		assert(func != nullptr);
